Small improvements in incident dashboard.
integria.css: This fix the "buttons" generated in Incident dashboard
when search name is a bit larger and doesn't fit on a single line

incident_dashboard.php: We omit the count for each incident due it's
great impact in load speed. Just display the custom search name +link
to that search.